companyapi updated/companyapi/requirements.txt
Django>=5.2.4
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.1
PyJWT>=2.10.1
requests>=2.31.0
